move test fixtures for individual tunables one level down into tunables/ folder as they are not used anywhere else
